user auth

+++ controllers/stuff.js
@@ -1,30 +1,58 @@
-const {Thing} = require("../models/thing");
+const { Thing } = require("../models/thing");
 
 exports.deleteOneThing = (req, res) => {
   const id = req.params.id;
-  Thing.deleteOne({ _id: id })
-    .then(() => {
-      res.status(200).json({ message: "deleted Successfully" });
+  const userId = req.auth.userId;
+  Thing.findById(id)
+    .then((thing) => {
+      if (thing && thing.userId != userId) {
+        return res
+          .status(401)
+          .json({ message: "U dont hav e the right to delete this item" });
+      }
+      Thing.deleteOne({ _id: id })
+        .then((state) => {
+          if (!state.deletedCount) {
+            return res.status(200).json({ message: "item already deleted" });
+          }
+          return res.status(200).json({ message: "deleted Successfully" });
+        })
+        .catch((error) => {
+          return res.status(400).json({ error });
+        });
     })
-    .catch((error) => {
-      res.status(400).json({ error });
-    });
+    .catch((error) => res.status(401).json({ error }));
 };
 
 exports.updateOneThing = (req, res, next) => {
-  const updatedThing = new Thing({
-    _id: req.params.id,
-    title: req.body.title,
-    description: req.body.description,
-    imageUrl: req.body.imageUrl,
-    price: req.body.price,
-    userId: req.body.userId,
-  });
-  Thing.updateOne({ _id: req.params.id }, updatedThing)
-    .then(() => {
-      res.status(201).json({
-        message: "Thing updated successfully!",
+  const userId = req.auth.userId;
+  const id = req.params.id;
+  Thing.findById(id)
+    .then((thing) => {
+      if (thing && thing.userId != userId) {
+        return res
+          .status(401)
+          .json({ message: "U dont hav e the right to update this item" });
+      }
+      const updatedThing = new Thing({
+        _id: req.params.id,
+        title: req.body.title,
+        description: req.body.description,
+        imageUrl: req.body.imageUrl,
+        price: req.body.price,
+        userId: req.body.userId,
       });
+      Thing.updateOne({ _id: req.params.id }, updatedThing)
+        .then(() => {
+          res.status(201).json({
+            message: "Thing updated successfully!",
+          });
+        })
+        .catch((error) => {
+          res.status(400).json({
+            error: error,
+          });
+        });
     })
     .catch((error) => {
       res.status(400).json({
@@ -34,7 +62,8 @@ exports.updateOneThing = (req, res, next) => {
 };
 
 exports.getAllThings = (req, res, next) => {
-  Thing.find().then((things) => {
+  Thing.find()
+    .then((things) => {
       res.status(200).json(things);
     })
     .catch((error) => res.status(400).json({ error }));
